requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
boto3==1.28.23
botocore==1.31.54
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.6
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.1
faiss-cpu==1.8.0
fastapi==0.110.3
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.34.1
inference-cpu==0.11.2
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain-core==0.2.3
langchain-openai==0.1.8
langgraph==0.0.61
langsmith==0.1.67
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.15.1
openai==1.30.5
opencv-python==4.8.0.76
opencv-python-headless==4.10.0.82
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.3.0
pluggy==1.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pybase64==1.3.2
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydot==2.0.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rf-clip==1.0
rich==13.5.2
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.6.2
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
skypilot==0.5.0
sniffio==1.3.1
starlette==0.37.2
structlog==24.2.0
supervision==0.20.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
tifffile==2024.5.22
tiktoken==0.7.0
time-machine==2.14.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.28
ultralytics-thop==0.2.7
uritemplate==4.1.1
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.9.4
zxing-cpp==2.2.0
